+++ package.js
@@ -16,7 +16,10 @@ Package.onUse( function( api ){
         'MD_BUTTON_SAVE',
         'MD_BUTTON_YES',
         'MD_BUTTON_NO',
-        'MD_POSITION_CENTER'
+        'MD_POSITION_CENTER',
+        'MD_VERBOSE_NONE',
+        'MD_VERBOSE_CONFIGURE',
+        'MD_VERBOSE_NOMODAL'
     ]);
     api.mainModule( 'src/client/js/index.js', 'client' );
     api.mainModule( 'src/server/js/index.js', 'server' );