+++ src/client/js/constants.js
@@ -19,5 +19,9 @@ MD_BUTTON_YES = 'MD_BUTTON_YES';
 MD_BUTTON_NO = 'MD_BUTTON_NO';
 
 // vertical position
-
 MD_POSITION_CENTER = 'MD_POSITION_CENTER';
+
+// verbosity level
+MD_VERBOSE_NONE           = 0x00;
+MD_VERBOSE_CONFIGURE      = 0x01 <<  0;
+MD_VERBOSE_NOMODAL        = 0x01 <<  1;