+++ src/common/js/config.js
@@ -1,8 +1,0 @@
-/*
- * pwix:modal/src/common/js/config.js
- */
-
-//console.log( 'pwix:modal/src/common/js/config.js defining globally exported pwixModal object' );
-
-pwixModal = {
-};