+++ src/common/js/global_export.js
@@ -1,0 +1,6 @@
+/*
+ * pwix:modal/src/common/js/global_export.js
+ */
+
+pwixModal = {
+};