+++ src/common/js/index.js
@@ -2,5 +2,7 @@
  * pwix:modal/src/common/js/index.js
  */
 
-import './config.js';
+import './global_export.js';
+//
+import './configure.js';
 import './i18n.js';